enhance the media details pager indicator
